job_scheduler/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
channels>=4.0.0
django-crispy-forms>=2.0
python-dotenv>=1.0.0
celery>=5.3.0
redis>=5.0.0
psycopg2-binary>=2.9.0
gunicorn>=21.0.0
crispy-bootstrap4